Lesson 23 - Updating Goals List
This creates a state to manage the goals list and then updates that state whenever the button is pressed that calls the 'addGoalHandler' function. It then dynamically updates the list by going through the array by mapping the various goals and then outputting the component to display the individual goal.