Lesson 25 - Adding a ScrollView
This adds a ScrollView to the goals list in case the list exceeds the viewable area. To maintain the layout ratio, the ScrollView needs to be placed inside the `goalsContainer` View instead of replacing it. Also, the `alwaysBounceVertical` is iOS specific and stops the bounce/scrollability if the list does not exceed the view port. See here for more info: https://reactnative.dev/docs/scrollview